+++ .gitignore
@@ -1,2 +1,3 @@
 __pycache__/game.cpython-313.pyc
 __pycache__/rrt_star.cpython-313.pyc
+__pycache__/game.cpython-312.pyc
